requirements.txt
flask
gunicorn
google-cloud-datastore
google-cloud-secret-manager